# swagger.yaml
openapi: 3.0.0
info:
  title: Arvog API Documentation
  description: API documentation for User, Category, Product & Bulk Upload Management
  version: 1.0.0
servers:
  - url: http://localhost:5000/api
    description: Local API Server
paths:
...
  /bulk-status/{jobId}:
    get:
      summary: Check bulk job status
      tags: [Bulk Upload]
      parameters:
        - in: path
          name: jobId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Status returned